Functional App Build!
Looks like it was a few different issues. The native build compilation error with Sharp was as a dependency of favicons-webpack-plugin, which looks to be something like PWA support. I can add that on my own as a standalone thing later on, so I disabled that for the time being.

Once that was fixed, Webpack wouldn't run properly, and it turned out to be what seems like a deprecated Node feature of SSL which was removed, so you have to pass a specific flag to Node to enable it again.

Going to try seeing how the built version of the site looks, and I want to try moving things over to Vite too, for modern bundling tech.

https://github.com/webpack/webpack/issues/14532
https://unix.stackexchange.com/questions/381902/permission-denied-when-we-use-npm-start-using-webpack
https://stackoverflow.com/questions/1158091/defining-a-variable-with-or-without-export
https://dev.to/wojtekmaj/migrating-your-react-app-from-webpack-to-vite-inp