Vite Node Polyfills Attempt
Not quite sure yet how to properly polyfill these with Vite. This package looks like an alternative, but it's not fully fixing the issues yet.

This commit's changes didn't really help with it, this is more where I was trying to go, it needs an actual setup to get things working, maybe go back to the changes before this one.

https://stackoverflow.com/questions/76431747/module-has-been-externalized-for-browser-compatibility-error-in-vite-build
https://www.npmjs.com/package/vite-plugin-node-polyfills